+++ R/zzz.R
@@ -1,0 +1,7 @@
+.onLoad <- function(libname, pkgname) {
+    tools::vignetteEngine("rasciidoc", package=pkgname,
+                          pattern="[.][Rr]asciidoc$",
+                          weave=vweave,
+                          tangle=vtangle
+                          )
+}